+++ tests/integration/sugar/test_account.py
@@ -1,7 +1,7 @@
 from tests.integration.integration_test_case import IntegrationTestCase
 from tests.integration.it_utils import (
     FUNDING_AMOUNT,
-    fund_wallet_sync,
+    fund_wallet,
     sign_and_reliable_submission_async,
     test_async_and_sync,
 )
@@ -13,7 +13,7 @@ from xrpl.models.transactions import Payment
 from xrpl.wallet import Wallet
 
 NEW_WALLET = Wallet.create()
-fund_wallet_sync(NEW_WALLET)
+fund_wallet(NEW_WALLET)
 EMPTY_WALLET = Wallet.create()